Organize by moving component and section folders
