+++ src/app/page.tsx
@@ -1,4 +1,4 @@
-import { HeroSection } from "./components/sections/hero-section";
+import { HeroSection } from "./sections/hero-section";
 
 export default function Home() {
   return (
