+++ src/app/sections/hero-section.tsx
