+++ src/components/logo-component.tsx
